ステータス部分の作成を完了

+++ Assets/Mansion/Scripts/Main/CountLabelController.cs
@@ -5,7 +5,7 @@ public class CountLabelController : MonoBehaviour {
 
 	public UILabel countLabel;
 
-	void SetCount(int count){
+	void SetCount(long count){
 		countLabel.text = count.ToString();
 	}
